New connectSqlite3Ext function
This allows prepared statements to stay open, so that one can avoid the
penalty of recompiling them each time.  Applications that disable auto
finish MUST finish() *every* statement handle they prepare().

This also supports use of native octet encodings of non-UTF-8 filenames.
The existing connect functions are re-implemented in terms of
connectSqlite3Ext.

When auto-finish is off, it makes no sense to track the open statement
handles that only the user will close.  Therefore, ChildList becomes
a (Maybe ChildList) and is unused when auto-finish is disabled.

Since we're not finishing queries when the last result row is read,
we at least promptly "reset" the statement, freeing up some of the
underlying resources.

Finally, single-use statements that the library prepares for its own
internal purposes are never auto-finished, and finished explicitly
instead.  Most of these will become cached statements in the next commit.

The previous implementation of clone() was incorrect, it forgot the
filename encoding, this is now preserved along with the auto-finish
flag.