Added tests for auto-finish off mode.
With implicit auto-finish of course the finish happens at the right time
automatically, but with auto-finish off, a bit more care is required to
not reset the query before all the desired rows are examined.

The explicit "finish" for the statements used to fetch all rows when
using the lazy fetchAllRows needs to be delayed until the rows have
been fully processed (by comparing with the expected values).  This
is not needed in the non-lazy fetchAllRows' case.

Replaced quickQuery with safeQuickQuery' when not explicitly testing
quickQuery.  The latter is not safe when auto-finish is off.  Applications
that disable auto-finish must avoid functions like quickQuery and quickQuery'
internally leak prepared statements.  NB: quickQuery' could be changed to
explicitly finish its statement handle.